update logging
In order to place all logs related to a specific node in the same folder,
I added another parameter in the join_hyperv function with the sole
purpose of specifying which node is being built (either hyperv01 or
hyperv02).
Also moved the creation of the directories for the windows/hyper-v logs in
the deploy_devstack script in order to also push the create-environment
logs directly where they belong. The folders are being made before
running devstack so it won't get in the way in case we'll build devstack
and the windows nodes in parallel.